complete wallet side nav
